+++ libraries/exoplayer/src/test/java/androidx/media3/exoplayer/source/ProgressiveMediaPeriodTest.java
@@ -39,6 +39,7 @@ import androidx.media3.exoplayer.trackselection.ExoTrackSelection;
 import androidx.media3.exoplayer.upstream.DefaultAllocator;
 import androidx.media3.exoplayer.upstream.DefaultLoadErrorHandlingPolicy;
 import androidx.media3.exoplayer.util.ReleasableExecutor;
+import androidx.media3.extractor.DefaultExtractorsFactory;
 import androidx.media3.extractor.Extractor;
 import androidx.media3.extractor.ExtractorsFactory;
 import androidx.media3.extractor.mp4.Mp4Extractor;
@@ -145,6 +146,60 @@ public final class ProgressiveMediaPeriodTest {
     mediaPeriod.release();
   }
 
+  @Test
+  public void getTrackGroups_multipleTracks_setsExpectedPrimaryTrackGroupId() throws Exception {
+    ProgressiveMediaPeriod mediaPeriod =
+        createMediaPeriod(Uri.parse("asset://android_asset/media/mp4/sample.mp4"));
+    TrackGroupArray trackGroups = mediaPeriod.getTrackGroups();
+
+    assertThat(trackGroups.length).isEqualTo(2);
+    assertThat(trackGroups.get(0).type).isEqualTo(C.TRACK_TYPE_VIDEO);
+    assertThat(trackGroups.get(1).type).isEqualTo(C.TRACK_TYPE_AUDIO);
+    assertThat(trackGroups.get(0).getFormat(0).primaryTrackGroupId).isNull();
+    assertThat(trackGroups.get(1).getFormat(0).primaryTrackGroupId)
+        .isEqualTo(trackGroups.get(0).id);
+
+    mediaPeriod =
+        createMediaPeriod(
+            Uri.parse("asset://android_asset/media/mp4/h265_with_metadata_track.mp4"));
+    trackGroups = mediaPeriod.getTrackGroups();
+
+    assertThat(trackGroups.length).isEqualTo(3);
+    assertThat(trackGroups.get(0).type).isEqualTo(C.TRACK_TYPE_METADATA);
+    assertThat(trackGroups.get(1).type).isEqualTo(C.TRACK_TYPE_AUDIO);
+    assertThat(trackGroups.get(2).type).isEqualTo(C.TRACK_TYPE_VIDEO);
+    assertThat(trackGroups.get(0).getFormat(0).primaryTrackGroupId)
+        .isEqualTo(trackGroups.get(2).id);
+    assertThat(trackGroups.get(1).getFormat(0).primaryTrackGroupId)
+        .isEqualTo(trackGroups.get(2).id);
+    assertThat(trackGroups.get(2).getFormat(0).primaryTrackGroupId).isNull();
+
+    mediaPeriod =
+        createMediaPeriod(Uri.parse("asset://android_asset/media/mp4/sample_with_vobsub.mp4"));
+    trackGroups = mediaPeriod.getTrackGroups();
+
+    assertThat(trackGroups.length).isEqualTo(3);
+    assertThat(trackGroups.get(0).type).isEqualTo(C.TRACK_TYPE_VIDEO);
+    assertThat(trackGroups.get(1).type).isEqualTo(C.TRACK_TYPE_AUDIO);
+    assertThat(trackGroups.get(2).type).isEqualTo(C.TRACK_TYPE_TEXT);
+    assertThat(trackGroups.get(0).getFormat(0).primaryTrackGroupId).isNull();
+    assertThat(trackGroups.get(1).getFormat(0).primaryTrackGroupId)
+        .isEqualTo(trackGroups.get(0).id);
+    assertThat(trackGroups.get(2).getFormat(0).primaryTrackGroupId)
+        .isEqualTo(trackGroups.get(0).id);
+
+    mediaPeriod =
+        createMediaPeriod(Uri.parse("asset://android_asset/media/ts/sample_with_id3.adts"));
+    trackGroups = mediaPeriod.getTrackGroups();
+
+    assertThat(trackGroups.length).isEqualTo(2);
+    assertThat(trackGroups.get(0).type).isEqualTo(C.TRACK_TYPE_AUDIO);
+    assertThat(trackGroups.get(1).type).isEqualTo(C.TRACK_TYPE_METADATA);
+    assertThat(trackGroups.get(0).getFormat(0).primaryTrackGroupId).isNull();
+    assertThat(trackGroups.get(1).getFormat(0).primaryTrackGroupId)
+        .isEqualTo(trackGroups.get(0).id);
+  }
+
   private static @SampleStream.ReadDataResult int readProgressiveStream(
       ProgressiveMediaPeriod mediaPeriod, int trackIndex, DecoderInputBuffer buffer) {
     return mediaPeriod.readData(trackIndex, new FormatHolder(), buffer, /* readFlags= */ 0);
@@ -156,13 +211,37 @@ public final class ProgressiveMediaPeriodTest {
       @Nullable Executor executor,
       @Nullable Consumer<Executor> executorReleased)
       throws TimeoutException {
+    return createMediaPeriod(
+        Uri.parse("asset://android_asset/media/mp4/sample.mp4"),
+        extractor,
+        imageDurationUs,
+        executor,
+        executorReleased);
+  }
+
+  private static ProgressiveMediaPeriod createMediaPeriod(Uri uri) throws TimeoutException {
+    return createMediaPeriod(
+        uri,
+        new BundledExtractorsAdapter(new DefaultExtractorsFactory()),
+        /* imageDurationUs= */ C.TIME_UNSET,
+        /* executor= */ null,
+        /* executorReleased= */ null);
+  }
+
+  private static ProgressiveMediaPeriod createMediaPeriod(
+      Uri uri,
+      ProgressiveMediaExtractor extractor,
+      long imageDurationUs,
+      @Nullable Executor executor,
+      @Nullable Consumer<Executor> executorReleased)
+      throws TimeoutException {
     AtomicBoolean sourceInfoRefreshCalled = new AtomicBoolean(false);
     ProgressiveMediaPeriod.Listener sourceInfoRefreshListener =
         (durationUs, seekMap, isLive) -> sourceInfoRefreshCalled.set(true);
     MediaPeriodId mediaPeriodId = new MediaPeriodId(/* periodUid= */ new Object());
     ProgressiveMediaPeriod mediaPeriod =
         new ProgressiveMediaPeriod(
-            Uri.parse("asset://android_asset/media/mp4/sample.mp4"),
+            uri,
             new AssetDataSource(ApplicationProvider.getApplicationContext()),
             extractor,
             DrmSessionManager.DRM_UNSUPPORTED,